reorganize plugins structure
