+++ src/plugins/index.ts
@@ -1,0 +1,3 @@
+export * from './driveBrowserPlugin';
+export * from './launcherPlugin';
+export * from './driveDialogPlugin';
